pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyqaunicore"
version = "1.0.0"
description = "Universal automating testing framework core for WEB-applications"
authors = [
    { name = "Dmitry Vorobjev" }
]
readme = "README.md"
license = {text = "NTC MEKHANOTRONIKA License"}
requires-python = ">=3.11"
dependencies = [
    "ansi2html>=1.9",
    "faker>=30.1",
    "loguru>=0.7",
    "psutil>=5.9",
    "pydantic>=2.7",
    "pytest-check>=2.3",
    "pytest-html>=4.1",
    "pytest-metadata>=2.0",
    "sqlmodel>=0.0",
    "tenacity>=8.2",
    "transliterate>=1.10",
    "websockets>=14.1",
]
classifiers = [
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Intended Audience :: Developers",
    "Intended Audience :: Testers",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv*",
]

# Same as Black.
line-length = 110
indent-width = 4

# Assume Python 3.11
target-version = "py311"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
extend-select = ["I"]
fixable = ["ALL"]